add options in predict.main() to predict on unseen text files
add input params for predict.main() for the directory containing files to be predicted on
and for common filename for those files.
dQ expects files to have the form test.src or test.mt and for them to be in the same dir
parameters['TEXT_FILES'] and 'DATA_ROOT_PATH' are modified for prediction
so that they refer to the specified files.