facilitate training on minimised metrics (ie rmse and mae)
this commit makes changes to the way the EarlyStopping callback from Keras Wrapper
is added to the callbacks for training. It is now added in deepquest.train.buildCallbacks
which allows us to set the want_to_minimize parameter dependent on our metrics
previously is was hard-set in keras_wrapper.extra.callbacks and was in effect, a bug.

Further, this commit removes the model save from deepquest.callbacks.EvalPerformance.evaluation
as saveModel is already called by keras_wrapper's on_epoch_end callback which was resulting
in the model being saved twice each epoch.